ops-platform/requirements.txt
django>=4.2,<5.0
psycopg2-binary>=2.9.9
djangorestframework>=3.15
django-environ>=0.11
anthropic>=0.25
requests>=2.31
celery>=5.3
redis>=5.0
django-celery-results>=2.5
gunicorn>=21.2
django-jazzmin>=3.0
